+++ lib/public/views/lhcPeriods/ActiveColumns/formatLhcPeriodNameWithLinksToAssociations.js
