+++ lib/public/views/lhcPeriods/ActiveColumns/lhcPeriodsActiveColumns.js
@@ -15,7 +15,7 @@ import { h } from '/js/src/index.js';
 import { formatDistinctLhcBeamEnergies } from '../format/formatDistinctLhcBeamEnergies.js';
 import { formatLhcPeriodYear } from '../format/formatYear.js';
 import { textFilter } from '../../../components/Filters/common/filters/textFilter.js';
-import { formatLhcPeriodNameWithLinksToAssociations } from '../format/formatLhcPeriodNameWithLinksToAssociations.js';
+import { formatLhcPeriodNameWithLinksToAssociations } from './formatLhcPeriodNameWithLinksToAssociations.js';
 
 /**
  * List of active columns for a generic periods table
